pyproject.toml
[build-system]
requires = ["setuptools>=42"]
build-backend = "setuptools.build_meta"

[project]
name = "eeganalyzer"
version = "0.1.0"
authors = [
    {name = "SoenkevL", email = "soenke.v.l@outlook.com"},
]
description = "A comprehensive tool for analyzing and visualizing EEG data with focus on customizable metric analysis"
readme = "README.md"
requires-python = ">=3.8"
classifiers = [
    "Programming Language :: Python :: 3",
    "License :: OSI Approved :: GPL-3.0-or-later",
    "Operating System :: OS Independent",
    "Topic :: Scientific/Engineering :: Medical Science Apps.",
    "Topic :: Scientific/Engineering :: Bio-Informatics",
]
dependencies = [
    "edgeofpy==0.0.1",
    "icecream==2.1.3",
    "customtkinter==5.2.2",
    "mat4py==0.6.0",
    "matplotlib==3.9.2",
    "mne==1.8.0",
    "mne-qt-browser==0.6.3",
    "multiprocesspandas==1.1.5",
    "neurokit2==0.2.10",
    "numpy==1.26.4",
    "pandas==2.2.3",
    "powerlaw==1.5",
    "PyYAML==6.0.2",
    "PyQt6==6.7.1",
    "scipy==1.14.1",
    "SQLAlchemy>=2.0.40",
]

[project.urls]
"Homepage" = "https://github.com/SoenkevL/EEGAnalyzer"
"Bug Tracker" = "https://github.com/SoenkevL/EEGAnalyzer/issues"

[project.scripts]
eeganalyzer = "eeganalyzer.cli.cli:main"
eegviewer = "gui.run_metrics_viewer:main"